Feat : Order 서비스 1차 구현

+++ src/main/java/com/konkuk/bit/bitpay/order/service/OrderService.java
@@ -2,7 +2,6 @@ package com.konkuk.bit.bitpay.order.service;
 
 import com.konkuk.bit.bitpay.order.domain.Order;
 import com.konkuk.bit.bitpay.order.dto.OrderCreateDto;
-import com.konkuk.bit.bitpay.order.dto.OrderUpdateDto;
 
 import java.util.List;
 import java.util.Optional;
@@ -19,6 +18,6 @@ public interface OrderService {
     boolean cancelOrder(Long orderId);
 
     // Order List
-    List<Order> getOrderListByTableNumber(Long tableNumber);
-    List<Order> getOrderListByTableNumberAndStatus(Long tableNumber, String status);
+    List<Order> getOrderListByTableNumber(Integer tableNumber);
+    List<Order> getOrderListByTableNumberAndStatus(Integer tableNumber, String status);
 }
